Update getLocalChain, update SyncBlockchain, refactor peer tests
